Fix Windows GPU Usage Issue (#252)

+++ Source/Core/Utils/ERS_CLASS_GPURequest/ERS_CLASS_GPURequest.h
@@ -11,7 +11,7 @@
 extern "C" {
 #endif
 
-__declspec(dllexport) DWORD NvOptimusEnablement = 1;
+__declspec(dllexport) unsigned long NvOptimusEnablement = 1;
 __declspec(dllexport) int AmdPowerXpressRequestHighPerformance = 1;
 
 #ifdef __cplusplus
